commands: move describe code to describe.rs
